+++ packages/react-ui/storybook/stories/widgetsUI/LegendWidgetUI.stories.js
@@ -66,6 +66,22 @@ const LegendNotFoundTemplate = () => {
   return <LegendWidgetUI layers={layers}></LegendWidgetUI>;
 };
 
+const LegendWithOpacityTemplate = () => {
+  const layers = [
+    {
+      id: 0,
+      title: 'Single Layer',
+      visible: true,
+      showOpacityControl: true,
+      opacity: 0.5,
+      legend: {
+        children: <div>Your Content</div>
+      }
+    }
+  ];
+  return <LegendWidgetUI layers={layers}></LegendWidgetUI>;
+};
+
 const LegendMultiTemplate = () => {
   const layers = [
     {
@@ -258,6 +274,7 @@ export const SingleLayer = LegendTemplate.bind({});
 export const MultiLayer = LegendMultiTemplate.bind({});
 export const MultiLayerCollapsed = LegendMultiTemplateCollapsed.bind({});
 export const NotFound = LegendNotFoundTemplate.bind({});
+export const LegendWithOpacityControl = LegendWithOpacityTemplate.bind({});
 export const Categories = LegendCategoriesTemplate.bind({});
 export const CategoriesAsStroke = LegendCategoriesStrokeTemplate.bind({});
 export const Icon = LegendIconTemplate.bind({});