+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Not released
 
+- Add legends tests/stories/types [#328](https://github.com/CartoDB/carto-react/pull/328)
+- Fix error in getPalette, causing wrong colors when using CARTOColors [#328](https://github.com/CartoDB/carto-react/pull/328)
+- Fix errors in LegendRamp with wrong labels [#328](https://github.com/CartoDB/carto-react/pull/328)
+- Dynamic tiling integration [#327](https://github.com/CartoDB/carto-react/pull/327)
 - Fix Switch input width [#323](https://github.com/CartoDB/carto-react/pull/323)
+- Add optional slider to modify layer opacity in Legend widget [#324](https://github.com/CartoDB/carto-react/pull/324)
 
 ## 1.2